Makes simulation.py work with all anchor coordinates
We previously hard coded Ax=Dx=Dy=0 right in the optimization script.

The new behavior is that the optimization will also look for suitable
Ax, Dx, and Dy coordinates in order to minimize the cost function.

This allows the xyz positions in which all the samples were collected
to be rotated compared to the Ax=Dx=Dy=0 one.

It will often be a motion tracking system that collects the xyz data,
and we don't know how well that system's coordinates are rotationally
aligned with Hangprinter's. We feel that it's better to find that out
with script rather than to try to force the user to do proper alignment
physically, before running this script.

The downside is that the user must collect ca 50% more data to make
the optimization converge. So we now require a minimum of
ca 6 sample points, up from ca 4 sample points.