Add pre-commit hook
At commit, will run:
-black
-isort
-pylint
-mypy
and a few more checks.

Only files changed in the commit will be checked. These static python code style checks are in quiet good harmony. If any discrepancies are found, it is recommended to let black decide how things should be and try to disable the other tools check. this is why pylint have many checks disabled at this point since it disagrees with black in some regards.

Also added VSCode workspace settings file with the matching configuration as pre-commit, so files should be automatically formatted and tested when saved in VSCode.